Main Routed
Added a NavigationService that will handle all routing. Make sure you define a new method when creating a new route and use these methods to route.

Added routing to the login that directs to the main landing page upon success and if they arrive at the login/signup page while already logged in.